% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\docType{data}
\name{Condition}
\alias{Condition}
\alias{Error}
\alias{Message}
\alias{SimpleCondition}
\alias{SimpleError}
\alias{SimpleMessage}
\alias{SimpleWarning}
\alias{Warning}
\title{Predefined condition classes}
\format{R6 class, with members \code{new} and \code{is}.}
\usage{
Condition

SimpleCondition

Error

SimpleError

Warning

SimpleWarning

Message

SimpleMessage
}
\description{
Each object represents a condition class.
}
\seealso{
\code{\link{ConditionClass}}
}
\keyword{datasets}

