% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\docType{data}
\name{ConditionClass}
\alias{ConditionClass}
\title{Generator class to create new condition classes}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConditionClass
}
\description{
Custom condition classes can be created with the help of this class.
}
\examples{
Oops <- ConditionClass$new("oops", Error)
Oops$new("Something went wrong")
}
\keyword{internal}

