% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedCondition.R
\docType{methods}
\name{FixedCondition}
\alias{FixedCondition}
\alias{FixedCondition,StQ,FixedConditionParam-method}
\alias{FixedCondition,StQList,FixedConditionParam-method}
\title{Method to stablish conditions for inteval-distance intervals}
\usage{
FixedCondition(object, Param)

\S4method{FixedCondition}{StQ,FixedConditionParam}(object, Param)

\S4method{FixedCondition}{StQList,FixedConditionParam}(object, Param)
}
\arguments{
\item{object}{Object of class \link{StQ} or \link{StQList} with data units and variables to set
the conditions.}

\item{Param}{Object of class \link{ConditionParam} with params to calculate condition values.
Depending on the type of condition.}
}
\value{
the same input \code{object} of class \link{StQ} or \link{StQList} including the specified
condition.
}
\description{
\code{FixedCondition} allows for establishing conditions in data sets for applying interval-distance edits.
}
\details{
This method checks which units meet the conditions specified in the input \code{Param} to apply
the validity control.
}
\examples{
\dontrun{
FixedCondition(object, Param)
}

}
\seealso{
\code{\link[Condition:FixedConditionParam-class]{FixedConditionParam}}
}
