% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenCondition.R
\docType{methods}
\name{GenCondition}
\alias{GenCondition}
\alias{GenCondition,StQ,GenConditionParam-method}
\alias{GenCondition,StQList,GenConditionParam-method}
\title{Method to stablish conditions for inteval-distance intervals}
\usage{
GenCondition(object, Param)

\S4method{GenCondition}{StQ,GenConditionParam}(object, Param)

\S4method{GenCondition}{StQList,GenConditionParam}(object, Param)
}
\arguments{
\item{object}{Object of class \link{StQ} or \link{StQList} with data units and variables to set
the conditions.}

\item{Param}{Object of class \link{ConditionParam} with params to calculate condition values.}
}
\value{
the same input \code{object} of class \link{StQ} or \link{StQList} including the specified
condition.
}
\description{
\code{GenCondition} allows for establishing conditions in data sets for applying interval-distance edits.
}
\details{
This method checks which units meet the conditions specified in the input \code{Param} to apply
the validity control.
}
\examples{
\dontrun{
GenCondition(object, Param)
}

}
\seealso{
\code{\link[Condition:GenConditionParam-class]{GenConditionParam}}
}
