% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenConditionParam-class.R
\docType{class}
\name{GenConditionParam-class}
\alias{GenConditionParam-class}
\title{Clase S4 GenConditionParam para los parámetros del método Condition}
\description{
Definición de la clase S4 \code{GenConditionParam} que contiene los parámetros
que utiliza el método \code{\link{Condition}}.
}
\details{
Esta clase define condiciones generales sobre las variables que deben verificarse
para que el control sea efectivo sobre la unidad.
}
\section{Slots}{

\describe{
\item{\code{EditName}}{\code{Vector} de tipo \code{integer} de longitud 1 con el nombre
del Edit para el que se establece la condición.}

\item{\code{ConditionExp}}{\code{vector} de tipo \code{character} con la expresión que
define la condición que se debe cumplir para que el control de validación
tenga efecto.}

\item{\code{by}}{\code{NULL} o \code{vector} de tipo \code{character} con los nombres
de las variables por las que queremos agrupar. Por defecto toma el valor
\code{NULL}.}

\item{\code{keyVar}}{\code{NULL} o \code{vector} de tipo \code{character} con los nombres
de las variables que se utilizan como clave. Por defecto toma el valor
\code{NULL}.}

\item{\code{lag}}{\code{NULL} o \code{vector} de tipo \code{integer} de longitud 1 con
el valor del retardo. Por defecto toma el valor \code{2L}.}

\item{\code{na.imp}}{\code{vector} de tipo \code{integer} de longitud 1, que indica si
los valores NA se imputan a 0L o a 1L.}

\item{\code{ConditionName}}{\code{Vector} de tipo \code{integer} de longitud 1 con el
nombre de la variable generada que contiene la condición.}
}}

\examples{
# Un prototipo vacío
new(Class = 'GenConditionParam')
# Ejemplo no vacío
new(Class = 'GenConditionParam',
    EditName = 'LCN_W_1',
    ConditionExp = 'is.na(ccaa)',
    keyVar = 'NOrden',
    ConditionName = 'Condicion')
}
