% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.condition}
\alias{is.condition}
\alias{is.error}
\alias{is.message}
\alias{is.warning}
\title{Check if the condition is of the appropriate type}
\usage{
is.condition(x, ...)

is.error(x, ...)

is.warning(x, ...)

is.message(x, ...)
}
\arguments{
\item{x}{The object to check}

\item{...}{Ignored}
}
\description{
Check if the condition is of the appropriate type
}

