% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSCondition.R
\docType{methods}
\name{TSCondition}
\alias{TSCondition}
\alias{TSCondition,StQList,TSConditionParam-method}
\title{Method to stablish conditions for inteval-distance intervals}
\usage{
TSCondition(object, Param)

\S4method{TSCondition}{StQList,TSConditionParam}(object, Param)
}
\arguments{
\item{object}{Object of class \link{StQ} or \link{StQList} with data units and variables to set
the conditions.}

\item{Param}{Object of class \link{TSConditionParam} with params to calculate condition values.}
}
\value{
object of class \code{StQ} with the following columns: the columns needed to identify the units
in the input \code{object} and a columns names 'Condition' with values 1L y 0L depending on
whether unit satisfies the condition or not.
}
\description{
\code{TSCondition} allows for establishing conditions in data sets for applying interval-distance edits.
}
\details{
This method checks which units meet the conditions specified in the input \code{Param} to apply
the validity control.
}
\examples{
\dontrun{
TSCondition(object, Param)
}

}
\seealso{
\code{\link[Condition:TSConditionParam-class]{TSConditionParam}}
}
