% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSConditionParam-class.R
\docType{class}
\name{TSConditionParam-class}
\alias{TSConditionParam-class}
\title{Clase S4 TSConditionParam para los parámetros del método Condition}
\description{
Definición de la clase S4 \code{TSConditionParam} que contiene los parámetros
que utiliza el método \code{\link{Condition}}.
}
\details{
Esta clase define condiciones sobre la serie temporal de las variables que
deben verificarse para que el control sea efectivo sobre la unidad.
}
\section{Slots}{

\describe{
\item{\code{EditName}}{\code{Vector} de tipo \code{integer} de longitud 1 con el nombre
del Edit para el que se establece la condición.}

\item{\code{VarName}}{\code{Vector} de tipo \code{integer} de longitud 1 con el nombre
de la variable sobre la que se establece la condición.}

\item{\code{NPeriods}}{\code{Vector} de tipo \code{integer} de longitud 1 con el número
de periodos hacia atrás con los que vamos a calcular la mediana.}

\item{\code{na.imp}}{\code{vector} de tipo \code{integer} de longitud 1, que indica si
los valores NA se imputan a 0L o a 1L.}

\item{\code{ConditionName}}{\code{Vector} de tipo \code{integer} de longitud 1 con el
nombre de la variable generada que contiene la condición.}
}}

\examples{
# Un prototipo vacío
new(Class = 'TSConditionParam')
# Ejemplo no vacío
new(Class = 'TSConditionParam',
    EditName = 'EMPREMFIJ_W_1',
    NPeriods = 3L,
    VarName = 'c121',
    ConditionName = 'Condicion')
}
