% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo.R
\docType{data}
\name{Condition}
\alias{Condition}
\alias{Error}
\alias{Message}
\alias{SimpleCondition}
\alias{SimpleError}
\alias{SimpleMessage}
\alias{SimpleWarning}
\alias{Warning}
\title{Predefined condition classes}
\format{An object of class \code{ConditionClassBase} (inherits from \code{ConditionClass}, \code{R6}) of length 7.}
\usage{
Condition

SimpleCondition

Error

SimpleError

Warning

SimpleWarning

Message

SimpleMessage
}
\description{
Condition classes, with members \code{new} and \code{is}.
}
\seealso{
\code{\link{ConditionClass}}
}
\keyword{datasets}

